# backend/schemas/common.py
from enum import Enum


class RoomStatus(str, Enum):
    VACANT = "VACANT"
    RESERVED = "RESERVED"
    OCCUPIED = "OCCUPIED"


class ReservationStatus(str, Enum):
    RESERVED = "RESERVED"
    CHECKED_IN = "CHECKED_IN"
    CHECKED_OUT = "CHECKED_OUT"
    CANCELLED = "CANCELLED"


class ErrorCode(str, Enum):
    ROOM_NOT_AVAILABLE = "ROOM_NOT_AVAILABLE"
    RESERVATION_CONFLICT = "RESERVATION_CONFLICT"
    RESERVATION_NOT_FOUND = "RESERVATION_NOT_FOUND"
    INVALID_STATE_TRANSITION = "INVALID_STATE_TRANSITION"